+++ PCA.h
@@ -25,6 +25,7 @@
 using namespace std;
 
 #include "H3DMath.h"
+#include "def.h"
 
 #define PCA_EPS            0.00000001    //PCA calucation accuracy
 #define PCA_ITERATION      60          //iteration times
@@ -36,25 +37,25 @@ class PCA
 public:
     PCA();
     ~PCA();
-	void dimReduc(double *mat,const int row,const int col,int &dim,double *P);//reduce dim
-	void dimIncrs(double *mat,const int row,const int dim,const int col,double *P);//increase dim
-	void getCovaMat(const double *mat,const int row,const int col,double *covMat);
+	void dimReduc(FLOAT *mat,const int row,const int col,int &dim,FLOAT *P);//reduce dim
+	void dimIncrs(FLOAT *mat,const int row,const int dim,const int col,FLOAT *P);//increase dim
+	void getCovaMat(const FLOAT *mat,const int row,const int col,FLOAT *covMat);
 
 private:
 
 
 	//calculate Eigenvector
-    int _calEigenVector(double CovMatrix[], int n, double Eigen[], double EigenVector[]);
-    void _Householder_Tri_Symetry_Diagonal(double a[], int n, double q[], double b[], double c[]);
-    int _Tri_Symmetry_Diagonal_Eigenvector(int n, double b[], double c[], double q[], double eps, int l);
+    int _calEigenVector(FLOAT CovMatrix[], int n, FLOAT Eigen[], FLOAT EigenVector[]);
+    void _Householder_Tri_Symetry_Diagonal(FLOAT a[], int n, FLOAT q[], FLOAT b[], FLOAT c[]);
+    int _Tri_Symmetry_Diagonal_Eigenvector(int n, FLOAT b[], FLOAT c[], FLOAT q[], FLOAT eps, int l);
 
 	//sort
 	void _mySwap(int *a,int *b);
-    void _quickSort(int left,int right,double array[],int id[]);
+    void _quickSort(int left,int right,FLOAT array[],int id[]);
 
 	//matrix operation
-	void _matrixMulti(double *rs,const double *a,const double *b,const int r,const int s,const int t);
-	void _matrixTransposition(double *m,const int row,const int col);
+	void _matrixMulti(FLOAT *rs,const FLOAT *a,const FLOAT *b,const int r,const int s,const int t);
+	void _matrixTransposition(FLOAT *m,const int row,const int col);
 };
 
 #endif
